requirements.txt
streamlit==1.29.0
google-generativeai==0.3.2
Pillow==10.1.0
pdf2image==1.16.3
pypdf==3.17.1
pandas==2.1.3
openpyxl==3.1.2
fpdf2==2.7.6
python-dotenv==1.0.0

